// src/API/AutoShopApi/models/AllUsersDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AutoShopAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserDTO } from './UserDTO';
import {
    UserDTOFromJSON,
    UserDTOFromJSONTyped,
    UserDTOToJSON,
    UserDTOToJSONTyped,
} from './UserDTO';

/**
 * 
 * @export
 * @interface AllUsersDTO
 */
export interface AllUsersDTO {
    /**
     * 
     * @type {Array<UserDTO>}
     * @memberof AllUsersDTO
     */
    users?: Array<UserDTO> | null;
    /**
     * 
     * @type {number}
     * @memberof AllUsersDTO
     */
    totalCount?: number;
}

/**
 * Check if a given object implements the AllUsersDTO interface.
 */
export function instanceOfAllUsersDTO(value: object): value is AllUsersDTO {
    return true;
}

export function AllUsersDTOFromJSON(json: any): AllUsersDTO {
    return AllUsersDTOFromJSONTyped(json, false);
}

export function AllUsersDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): AllUsersDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'users': json['users'] == null ? undefined : ((json['users'] as Array<any>).map(UserDTOFromJSON)),
        'totalCount': json['totalCount'] == null ? undefined : json['totalCount'],
    };
}

export function AllUsersDTOToJSON(json: any): AllUsersDTO {
    return AllUsersDTOToJSONTyped(json, false);
}

export function AllUsersDTOToJSONTyped(value?: AllUsersDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'users': value['users'] == null ? undefined : ((value['users'] as Array<any>).map(UserDTOToJSON)),
        'totalCount': value['totalCount'],
    };
}